refactor: move listWalletsByAccountId under app
